refactor(types): rename the integer type family
Rename the belt integer primitives so the sized names read as their C-like
widths and the arbitrary-precision pair takes the n-prefix:

  int/uint (Bits:0)  -> nint/nuint
  int8/uint8         -> sbyte/byte
  int16/uint16       -> short/ushort
  int32/uint32       -> int/uint
  int64/uint64       -> long/ulong

The semantics are unchanged: only the names move. int32 becoming int and int
becoming nint collide, so every identifier was renamed in a single dictionary
pass (each token looked up exactly once) rather than sequential substitution.

.belt sources were renamed on identifier boundaries; Go sources only inside
string literals, so Go's own int8/int32/int64 type usage is untouched and the
belt type-name strings (registry integerSpecs, ir.Builtin{Name:...}, the
default-int literal type, the LSP enum-base list, test fixtures and expected
messages) flip to the new names. The snumeric/unumeric unions are dropped: they
had no users beyond a prelude alias check.

The prelude int.belt/uint.belt declared a whole family each, so they are
renamed to the neutral signed.belt/unsigned.belt and the barrel's use paths
follow. Hardcoded byte offsets and LSP token positions over fixtures whose
type-name widths shifted were adjusted by hand; snapshots were regenerated with
-update (type-name text and its alignment markers only). Generated catalogs and
the editor grammar are unchanged — integer type names are not keywords.

+++ pkg/masterbelt/builtin/native.go
@@ -161,7 +161,7 @@ func stringMethods() []*ir.Method {
 var (
 	datetimeType ir.Type = &ir.Builtin{Name: "datetime"}
 	durationType ir.Type = &ir.Builtin{Name: "duration"}
-	intType      ir.Type = &ir.Builtin{Name: "int"}
+	intType      ir.Type = &ir.Builtin{Name: "nint"}
 )
 
 // comparisonMethods is the equality and ordering signature set shared by the
